migrate to autotools - refactor the code to be able to use autotools     * the actual logic remained unchanged, it was just moved to       different places - the original Makefile is staying for backwards-compatibility reasons
