+++ .gitignore
@@ -1,3 +1,16 @@
 main
 out.txt
 build/
+.cache/
+compile_commands.json
+autom4te.cache/
+aclocal.m4
+compile
+config.guess
+config.sub
+configure
+depcomp
+install-sh
+Makefile.in
+missing
+configure~
