+++ autogen.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+set -xe
+
+autoreconf -vi
+mkdir -p build
+cd build
+../configure
