+++ cleanup.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+rm -rf autom4te.cache build
+rm -f aclocal.m4 compile config.guess config.sub configure depcomp install-sh Makefile.in missing configure~ ar-lib ltmain.sh
